requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose==3.3.0
cryptography==41.0.7
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.0
reportlab==4.0.7
